// src/model/IUser.ts
export class IUser {
    _id: any = undefined;
    flagUpdate: boolean = false;
    isUser: boolean = undefined;
    isEnabled: boolean = undefined;
    user_range: string = undefined;
    user_sex: string = undefined;
    study: any = undefined;
    user_info: any = undefined;
    password: string = undefined;
    passwordHash: string = undefined;
    
    constructor(){}
}